AutonomousStoneArm + autonomous reorganization
Module for a stone arm that is used just in autonomous period.
