+++ SkyStone_RO060_app/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/AutonomousMovementTest.java
@@ -3,27 +3,32 @@ package org.firstinspires.ftc.teamcode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
+import org.firstinspires.ftc.teamcode.Modules.Auto_StoneArm;
 import org.firstinspires.ftc.teamcode.Modules.FoundationServos;
 import org.firstinspires.ftc.teamcode.Robot;
 
 @Autonomous
-@Disabled
-@Deprecated
 public class AutonomousMovementTest extends LinearOpMode {
 
     private Robot rb;
+    private Auto_StoneArm stoneArm = new Auto_StoneArm();
 
     public void runOpMode() throws InterruptedException{
         initRobot();
-int time=300;
+        int time=300;
         while(!isStarted()) idle();
+
+        /*while(opModeIsActive()){
+            stoneArm.Update(gamepad1);
+        }*/
         sleep(time);
-        rb.movement.rotateIMUAbsolute((float)Math.PI/2, 1f, this);
+        rb.movement.moveCM((float)Math.PI/2, 100, 1f, this);
         sleep(time);
-        rb.movement.rotateIMUAbsolute((float)Math.PI, 1f, this);
+        rb.movement.rotateIMURelative(Math.PI, 1f, this);
         sleep(time);
-        rb.movement.rotateIMUAbsolute((float)Math.PI*3/2f, 1f, this);
+        rb.movement.moveCM((float)Math.PI/2, 100, 1f, this);
         sleep(time);
         rb.movement.rotateIMUAbsolute(0, 1f, this);
         sleep(time);
