+++ SkyStone_RO060_app/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OldAutonomous/Parking_Right.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.OldAutonomous;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
@@ -9,7 +9,7 @@ import org.firstinspires.ftc.teamcode.Robot;
 @Autonomous
 @Disabled
 @Deprecated
-public class Parking_Left extends LinearOpMode {
+public class Parking_Right extends LinearOpMode {
 
     public Robot rb;
 
@@ -23,9 +23,9 @@ public class Parking_Left extends LinearOpMode {
             idle();
         }
 
-        rb.movement.moveCM((float)Math.PI/3, 100, 1f, this);
+        rb.movement.moveCM(2*(float)Math.PI/3, 100, 1f, this);
         sleep(100);
-        rb.movement.moveCM((float)Math.PI, 50, 1f, this);
+        rb.movement.moveCM(0, 50, 1f, this);
         sleep(100);
 
     }
