+++ SkyStone_RO060_app/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OldAutonomous/Parking_RoboKey.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.OldAutonomous;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
@@ -7,14 +7,15 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import org.firstinspires.ftc.teamcode.Robot;
 @Disabled
 @Deprecated
-@Autonomous (name="RED - STONES")
-public class Parking_IDL extends LinearOpMode {
+@Autonomous(name="BLUE - STONES")
+public class Parking_RoboKey extends LinearOpMode {
 
     public Robot rb;
 
 
 
     @Override
+
     public void runOpMode() throws InterruptedException {
         rb = new Robot(hardwareMap, this); //Initialization of the Robot
 
@@ -22,9 +23,9 @@ public class Parking_IDL extends LinearOpMode {
             idle();
         }
 
-        sleep(25000);
+        sleep(1000);
 
-        rb.movement.moveCM((float)Math.PI, 100, 1f, this);
+        rb.movement.moveCM(0, 100, 1f, this);
         sleep(50);
         rb.movement.moveCM((float)Math.PI/2, 100, 1f, this );
