+++ SkyStone_RO060_app/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OldAutonomous/RED_Parking_C.java
@@ -1,11 +1,11 @@
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.OldAutonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.Robot;
 
-@Autonomous (name = "BLUE Parking C- park near Center")
-public class BLUE_Parking_C extends LinearOpMode {
+@Autonomous (name = "RED Parking C- park near Center")
+public class RED_Parking_C extends LinearOpMode {
 
     public Robot rb;
 
@@ -16,9 +16,9 @@ public class BLUE_Parking_C extends LinearOpMode {
         initRobot();
 
         while(!isStarted()) idle();
-        rb.movement.moveCM((float)Math.PI,85,1f,this);
+        rb.movement.moveCM((float)Math.PI,-95,1f,this);
         sleep(time);
-        rb.movement.moveCM((float)Math.PI/2.2f, -80 , 1f, this);
+        rb.movement.moveCM((float)Math.PI/1.8f, -80 , 1f, this);
         sleep(time);
     }
