+++ SkyStone_RO060_app/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot.java
@@ -5,6 +5,7 @@ import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
+import org.firstinspires.ftc.teamcode.Modules.Auto_StoneArm;
 import org.firstinspires.ftc.teamcode.Modules.FoundationServos;
 import org.firstinspires.ftc.teamcode.Modules.MarkerArm;
 import org.firstinspires.ftc.teamcode.Modules.Movement;
@@ -17,6 +18,7 @@ public class Robot {
     public PrindereCub prindereCub;
     public MarkerArm markerArm;
     public FoundationServos foundationServos;
+    public Auto_StoneArm stoneArm;
 
     public Robot(HardwareMap hwm, LinearOpMode op) {
         //brat.Init("Brat", hwm);
@@ -34,6 +36,9 @@ public class Robot {
 
         foundationServos = new FoundationServos();
         foundationServos.Init("FoundationLeft", "FoundationRight", hwm);
+
+        stoneArm = new Auto_StoneArm();
+        stoneArm.Init("grabber", "arm", hwm);
     }
 
     public void LinearUpdate(Gamepad gamepad1, Gamepad gamepad2, OpMode op) {
